feat (provider): insert all files in index
