+++ .github/dependabot.yaml
@@ -1,8 +1,8 @@
 version: 2
 updates:
   - package-ecosystem: npm
-    directory: "/"
+    directory: '/'
     schedule:
       interval: weekly
-      time: "08:00"
+      time: '08:00'
     open-pull-requests-limit: 3
