+++ src/globals.d.ts
@@ -1,4 +1,12 @@
+type CreateReporterProps = {
+  reporter: string[];
+  reportsDirectory: string;
+  tempDirectory: string;
+};
+
 declare module 'c8/lib/report.js' {
-  const createReporter: any;
+  const createReporter: (props: CreateReporterProps) => {
+    run: () => Promise<void>;
+  };
   export default createReporter;
 }
